backend/pyproject.toml
[project]
name = "climbers-journal-backend"
version = "0.1.0"
description = "Climbers Journal Backend API"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sqlmodel>=0.0.22",
    "alembic>=1.14.0",
    "asyncpg>=0.30.0",
    "pydantic-settings>=2.7.0",
    "openai>=1.60.0",
    "httpx>=0.28.0",
    "python-multipart>=0.0.20",
    "Pillow>=10.0.0",
    "python-jose[cryptography]>=3.3.0",
    "arize-phoenix-otel>=0.1.0",
    "openinference-instrumentation-openai>=0.1.0",
    "opentelemetry-exporter-otlp-proto-grpc>=1.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.28.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
